requirements.txt
gradio>=4.0.0
torch>=2.1.0
torchvision
diffusers>=0.25.0
transformers>=4.38.0
accelerate>=0.27.0
opencv-python
pillow
omegaconf
numpy<2.0
sentencepiece
protobuf
scipy
timm
basicsr
realesrgan
ultralytics>=8.1.0
